+++ test/postgresql/full_test.sql
@@ -57,6 +57,7 @@
 \ir 47_row_filter_advanced.sql
 \ir 48_row_filter_multi_table.sql
 \ir 49_row_filter_prefill.sql
+\ir 50_block_lww_existing_data.sql
 
 -- 'Test summary'
 \echo '\nTest summary:'